aloh-langserv/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cymem==2.0.5
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
murmurhash==1.0.5
numpy==1.21.1
packaging==21.0
pathy==0.6.0
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
tqdm==4.62.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
yarl==1.6.3
